django/03_django_rest_framework/crudproject/crudapp/serializers.py
from rest_framework import serializers
from .models import Orders

class OrdersSerializer(serializers.ModelSerializer):
    class Meta:
        model = Orders
        fields = ['oid', 'fname', 'lname', 'price', 'mail', 'addr']