django/03_django_rest_framework/crudproject/requirements.txt
asgiref
Django
python-decouple
sqlparse
typing_extensions
tzdata
django-crispy-forms
crispy_bootstrap5
mysqlclient
djangorestframework